manul validate on delete

+++ SchoolSystem.API/Domain/Commands/Students/DeleteStudentCommandHandler.cs
@@ -1,25 +1,40 @@
-﻿using MediatR;
+﻿using FluentValidation;
+using MediatR;
+using SchoolSystem.API.Domain.Communication.Response;
 using SchoolSystem.API.Domain.Models.Students;
 using SchoolSystem.API.Domain.Repositories;
 
 namespace SchoolSystem.API.Domain.Commands.Students
 {
-    public class DeleteStudentCommandHandler : IRequestHandler<DeleteStudentCommand, Student>
+    public class DeleteStudentCommandHandler : IRequestHandler<DeleteStudentCommand,SchoolResponse<Student>>
     {
         private readonly IStudentsRepository studentsRepository;
+        private readonly IValidator<DeleteStudentCommand> validator;
 
-        public DeleteStudentCommandHandler(IStudentsRepository studentsRepository)
+        public DeleteStudentCommandHandler(IStudentsRepository studentsRepository
+            , IValidator<DeleteStudentCommand> validator)
         {
             this.studentsRepository = studentsRepository;
+            this.validator = validator;
         }
-        public async Task<Student> Handle(DeleteStudentCommand request, CancellationToken cancellationToken)
+        public async Task<SchoolResponse<Student>> Handle(DeleteStudentCommand request, CancellationToken cancellationToken)
         {
+            var resukt = this.validator.Validate(request);
+            if(!resukt.IsValid)
+            {
+                ErrorResponse <Student>errorResponse = new ErrorResponse<Student>();
+                foreach (var ex in resukt.Errors)
+                    errorResponse.AddError(ex.PropertyName, ex.ErrorMessage);
+                return errorResponse;
+
+            }
             Student deletedStudent = await this.studentsRepository.SelectOneAsync(request.Id);
             if (deletedStudent == null) { throw new Exception("Student Not found"); }
 
             await this.studentsRepository.DeleteAsync(deletedStudent);
-
-            return deletedStudent;
+            SuccessResponse<Student> successResponse = new SuccessResponse<Student>();
+            successResponse.Data = deletedStudent;
+            return successResponse;
         }
     }
 }
